Create an independent class XFEMCutElem
