+++ framework/include/XFEM/XFEM.h
@@ -15,23 +15,7 @@
 #ifndef XFEM_H
 #define XFEM_H
 
-// Local Includes -----------------------------------
-#include "libmesh/libmesh_common.h"
-#include "libmesh/libmesh.h" // libMesh::invalid_uint
-#include "libmesh/location_maps.h"
-#include "libmesh/mesh.h"
-#include "libmesh/elem.h"
-#include "libmesh/node.h"
-#include "libmesh/vector_value.h"
-#include "libmesh/point_locator_base.h"
-
-// C++ Includes   -----------------------------------
-#include <vector>
-
-#include "MaterialData.h"
-#include "ElementFragmentAlgorithm.h"
-
-using namespace libMesh;
+#include "XFEMCutElem.h"
 
 enum XFEM_CUTPLANE_QUANTITY
 {
@@ -44,30 +28,6 @@ enum XFEM_CUTPLANE_QUANTITY
 };
 
 class XFEM_geometric_cut;
-
-class XFEMCutElem
-{
-public:
-  XFEMCutElem(Elem* elem, const EFAelement * const CEMelem);
-  ~XFEMCutElem();
-
-private:
-
-  unsigned int _n_nodes;
-  std::vector<Node*> _nodes;
-  EFAelement _efa_elem;
-
-  Real _physical_volfrac;
-  Point get_node_coords(EFAnode* node, 
-                        MeshBase* displaced_mesh = NULL) const;
-
-public:
-  void calc_physical_volfrac();
-  Real get_physical_volfrac() const {return _physical_volfrac;}
-  Point get_origin(unsigned int plane_id, MeshBase* displaced_mesh=NULL) const;
-  Point get_normal(unsigned int plane_id, MeshBase* displaced_mesh=NULL) const;
-  const EFAelement * get_efa_elem() const;
-};
 
 /**
  * This is the \p XFEM class.  This class implements
